backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
alembic==1.13.3
httpx==0.27.2
firebase-admin==6.5.0
slowapi==0.1.9
structlog==24.4.0
pandas==2.2.2
openpyxl==3.1.5
sentry-sdk[fastapi]==2.8.0
cloudinary==1.40.0
sendgrid==6.11.0
python-dateutil==2.9.0
